Move Mongo Engine outside models
