+++ udata/factories.py
@@ -1,6 +1,6 @@
 import factory
 
-from .models import DateRange
+from .models import db
 
 
 class ModelFactory(factory.mongoengine.MongoEngineFactory):
@@ -14,7 +14,7 @@ class ModelFactory(factory.mongoengine.MongoEngineFactory):
 
 class DateRangeFactory(ModelFactory):
     class Meta:
-        model = DateRange
+        model = db.DateRange
 
     start = factory.Faker('date_between', start_date='-10y', end_date='-5y')
     end = factory.Faker('date_between', start_date='-5y', end_date='-2y')
