+++ udata/models/__init__.py
@@ -3,78 +3,14 @@ import warnings
 
 from urllib.parse import urlparse
 
-from bson import ObjectId, DBRef
-from flask_mongoengine import MongoEngine, MongoEngineSessionInterface
-from mongoengine.base import TopLevelDocumentMetaclass, get_document
 from mongoengine.errors import ValidationError
-from mongoengine.signals import pre_save, post_save
-
-from flask_storage.mongo import FileField, ImageField
 
 from udata import entrypoints
 from udata.errors import ConfigError
 
-from .badges_field import BadgesField
-from .taglist_field import TagListField
-from .datetime_fields import DateField, DateRange, Datetimed
-from .extras_fields import ExtrasField, OrganizationExtrasField
-from .slug_fields import SlugField
-from .url_field import URLField
-from .uuid_fields import AutoUUIDField
-from .owned import Owned, OwnedQuerySet
-from .queryset import UDataQuerySet
-from .document import UDataDocument, DomainModel
+from udata.mongo import *
 
 log = logging.getLogger(__name__)
-
-class UDataMongoEngine(MongoEngine):
-    '''Customized mongoengine with extra fields types and helpers'''
-    def __init__(self, app=None):
-        super(UDataMongoEngine, self).__init__(app)
-        self.BadgesField = BadgesField
-        self.TagListField = TagListField
-        self.DateField = DateField
-        self.Datetimed = Datetimed
-        self.ExtrasField = ExtrasField
-        self.OrganizationExtrasField = OrganizationExtrasField
-        self.SlugField = SlugField
-        self.AutoUUIDField = AutoUUIDField
-        self.Document = UDataDocument
-        self.DomainModel = DomainModel
-        self.DateRange = DateRange
-        self.BaseQuerySet = UDataQuerySet
-        self.BaseDocumentMetaclass = TopLevelDocumentMetaclass
-        self.FileField = FileField
-        self.ImageField = ImageField
-        self.URLField = URLField
-        self.ValidationError = ValidationError
-        self.ObjectId = ObjectId
-        self.DBRef = DBRef
-        self.Owned = Owned
-        self.OwnedQuerySet = OwnedQuerySet
-        self.post_save = post_save
-        self.pre_save = pre_save
-
-    def resolve_model(self, model):
-        '''
-        Resolve a model given a name or dict with `class` entry.
-
-        :raises ValueError: model specification is wrong or does not exists
-        '''
-        if not model:
-            raise ValueError('Unsupported model specifications')
-        if isinstance(model, str):
-            classname = model
-        elif isinstance(model, dict) and 'class' in model:
-            classname = model['class']
-        else:
-            raise ValueError('Unsupported model specifications')
-
-        try:
-            return get_document(classname)
-        except self.NotRegistered:
-            message = 'Model "{0}" does not exist'.format(classname)
-            raise ValueError(message)
 
 
 class FieldValidationError(ValidationError):
@@ -83,11 +19,6 @@ class FieldValidationError(ValidationError):
     def __init__(self, *args, field: str, **kwargs):
         self.field = field
         super().__init__(*args, **kwargs)
-
-
-db = UDataMongoEngine()
-session_interface = MongoEngineSessionInterface(db)
-
 
 # Load all core models and mixins
 from udata.core.spatial.models import *  # noqa
