+++ udata/mongo/__init__.py
@@ -1,0 +1,1 @@
+from .engine import db
