+++ udata/mongo/slug_fields.py
